@@ -1,5 +1,10 @@
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_ecommerce/common/widgets/brands/t_brand_card.dart';
+import 'package:flutter_ecommerce/common/widgets/shimmers/shimmer.dart';
+import 'package:flutter_ecommerce/features/shop/models/brand_model.dart';
+import 'package:flutter_ecommerce/features/shop/screens/brand/brand_products.dart';
+import 'package:get/get.dart';
 
 import '../../../utils/constants/colors.dart';
 import '../../../utils/constants/sizes.dart';
@@ -9,30 +14,34 @@ import '../custom_shapes/containers/rounded_container.dart';
 class TBrandShowCase extends StatelessWidget {
   const TBrandShowCase({
     super.key,
-    required this.images,
+    required this.images, required this.brand,
   });
 
+  final BrandModel brand;
   final List<String> images;
 
   @override
   Widget build(BuildContext context) {
-    return TRoudedContainer(
-      showBorder: true,
-      borderColor: TColors.darkGrey,
-      backgroundColor: Colors.transparent,
-      padding: const EdgeInsets.all(TSizes.md),
-      margin: const EdgeInsets.only(bottom: TSizes.spaceBtwItems),
-      child: Column(
-        children: [
-          /// brand with products count
-          const TBrandCard(showBorder: false,),
-          const SizedBox(height: TSizes.spaceBtwItems),
-
-          /// brand top 3 product images
-          Row(
-            children: images.map((image) => brandTopProductImageWidget(image, context)).toList(),
-          ),
-        ],
+    return InkWell(
+      onTap: () => Get.to(() => BrandProducts(brand: brand)),
+      child: TRoudedContainer(
+        showBorder: true,
+        borderColor: TColors.darkGrey,
+        backgroundColor: Colors.transparent,
+        padding: const EdgeInsets.all(TSizes.md),
+        margin: const EdgeInsets.only(bottom: TSizes.spaceBtwItems),
+        child: Column(
+          children: [
+            /// brand with products count
+            TBrandCard(showBorder: false, brand: brand),
+            const SizedBox(height: TSizes.spaceBtwItems),
+      
+            /// brand top 3 product images
+            Row(
+              children: images.map((image) => brandTopProductImageWidget(image, context)).toList(),
+            ),
+          ],
+        ),
       ),
     );
   }
@@ -44,7 +53,12 @@ class TBrandShowCase extends StatelessWidget {
         backgroundColor: THelperFunctions.isDarkMode(context) ? TColors.darkerGrey : TColors.light,
         margin: const EdgeInsets.only(right: TSizes.sm),
         padding: const EdgeInsets.all(TSizes.md),
-        child: Image(fit: BoxFit.contain, image: AssetImage(image)),
+        child: CachedNetworkImage(
+          fit: BoxFit.contain,
+          imageUrl: image,
+          progressIndicatorBuilder: (context, url, downloadProgress) => const TShimmerEffect(width: 100, height: 100),
+          errorWidget: (context, url, error) => const Icon(Icons.error),
+        ),
       ),
     );
   }
