@@ -18,16 +18,16 @@ class TChoiceChip extends StatelessWidget {
   Widget build(BuildContext context) {
     final isColor = THelperFunctions.getColor(text) != null;
     return ChoiceChip(
-      label: isColor ? SizedBox(): Text(text),
+      label: isColor ? const SizedBox(): Text(text),
       selected: selected,
       onSelected: onSelected,
       labelStyle: TextStyle(color: selected ? TColors.white : null),
       avatar: isColor ? TCircularContainer(width: 50, height: 50, backgroundColor: THelperFunctions.getColor(text)!):null,
-      shape: isColor ? CircleBorder():null,
-      labelPadding: isColor? EdgeInsets.all(0):null,
-      padding: isColor? EdgeInsets.all(0):null,
+      labelPadding: isColor ? const EdgeInsets.all(0) : null,
+      padding: isColor ? const EdgeInsets.all(0) : null,
+      shape: isColor ? const CircleBorder() : null,
+      backgroundColor: isColor ? THelperFunctions.getColor(text)! : null,
       selectedColor: isColor? Colors.green:null,
-      backgroundColor: isColor? Colors.green:null,
     );
   }
 }
