@@ -1,10 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_ecommerce/common/widgets/appbar/appbar.dart';
 import 'package:flutter_ecommerce/common/widgets/custom_shapes/containers/rounded_container.dart';
+import 'package:flutter_ecommerce/features/personalization/controllers/address_controller.dart';
 import 'package:flutter_ecommerce/features/personalization/screens/address/add_new_address.dart';
 import 'package:flutter_ecommerce/features/personalization/screens/address/widgets/single_address.dart';
 import 'package:flutter_ecommerce/utils/constants/colors.dart';
 import 'package:flutter_ecommerce/utils/constants/sizes.dart';
+import 'package:flutter_ecommerce/utils/helpers/cloud_helper_functions.dart';
 import 'package:get/get.dart';
 import 'package:iconsax/iconsax.dart';
 
@@ -13,6 +15,8 @@ class UserAddressScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final controller = Get.put(AddressController());
+
     return Scaffold(
       floatingActionButton: FloatingActionButton(
         onPressed: () => Get.to(() => const AddNewAddressScreen()),
@@ -24,14 +28,28 @@ class UserAddressScreen extends StatelessWidget {
         title:
             Text('Address', style: Theme.of(context).textTheme.headlineSmall),
       ),
-      body: const SingleChildScrollView(
+      body: SingleChildScrollView(
         child: Padding(
-          padding: EdgeInsets.all(TSizes.defaultSpace),
-          child: Column(
-            children: [
-              TSingleAddress(selectedAddress: true),
-              TSingleAddress(selectedAddress: false),
-            ],
+          padding: const EdgeInsets.all(TSizes.defaultSpace),
+          child: Obx(()=> FutureBuilder(
+              key: Key(controller.refreshData.value.toString()),
+              future: controller.getAllUserAddresses(),
+              builder: (context, snapshot) {
+                final response = TCloudHelperFunctions.checkMultiRecordState(snapshot: snapshot);
+                if (response != null) return response;
+            
+                final addresses = snapshot.data!;
+            
+                return ListView.builder(
+                  shrinkWrap: true,
+                  itemCount: addresses.length,
+                  itemBuilder: (_, index) => TSingleAddress(
+                      address: addresses[index],
+                      onTap: () => controller.selectAddress(addresses[index]),
+                  ),
+                );
+              }
+            ),
           ),
         ),
       ),
