@@ -1,63 +1,80 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_ecommerce/features/personalization/controllers/address_controller.dart';
+import 'package:flutter_ecommerce/features/personalization/models/address_model.dart';
 import 'package:flutter_ecommerce/utils/constants/colors.dart';
 import 'package:flutter_ecommerce/utils/constants/sizes.dart';
 import 'package:flutter_ecommerce/utils/helpers/helper_functions.dart';
+import 'package:get/get.dart';
 import 'package:iconsax/iconsax.dart';
 
 import '../../../../../common/widgets/custom_shapes/containers/rounded_container.dart';
 
 class TSingleAddress extends StatelessWidget {
-  const TSingleAddress({super.key, required this.selectedAddress});
+  const TSingleAddress({
+    super.key,
+    required this.address,
+    required this.onTap
+  });
 
-  final bool selectedAddress;
+  final AddressModel address;
+  final VoidCallback onTap;
 
   @override
   Widget build(BuildContext context) {
+    final controller = AddressController.instance;
     final dark = THelperFunctions.isDarkMode(context);
-    return TRoudedContainer(
-      padding: const EdgeInsets.all(TSizes.md),
-      width: double.infinity,
-      showBorder: true,
-      backgroundColor: selectedAddress
-          ? TColors.primary.withOpacity(0.5)
-          : Colors.transparent,
-      borderColor: selectedAddress
-          ? Colors.transparent
-          : dark
-              ? TColors.darkerGrey
-              : TColors.grey,
-      margin: const EdgeInsets.only(bottom: TSizes.spaceBtwItems),
-      child: Stack(
-        children: [
-          Positioned(
-            right: 5,
-            top: 0,
-            child: Icon(
-              selectedAddress ? Iconsax.tick_circle5 : null,
-              color: selectedAddress
-                  ? dark
-                      ? TColors.light
-                      : TColors.dark.withOpacity(0.2)
-                  : null,
+    return Obx(() {
+      final selectedAddressId = controller.selectedAddress.value.id;
+      final selectedAddress = selectedAddressId == address.id;
+        return InkWell(
+          onTap: onTap,
+          child: TRoudedContainer(
+            padding: const EdgeInsets.all(TSizes.md),
+            width: double.infinity,
+            showBorder: true,
+            backgroundColor: selectedAddress
+                ? TColors.primary.withOpacity(0.5)
+                : Colors.transparent,
+            borderColor: selectedAddress
+                ? Colors.transparent
+                : dark
+                ? TColors.darkerGrey
+                : TColors.grey,
+            margin: const EdgeInsets.only(bottom: TSizes.spaceBtwItems),
+            child: Stack(
+              children: [
+                Positioned(
+                  right: 5,
+                  top: 0,
+                  child: Icon(
+                    selectedAddress ? Iconsax.tick_circle5 : null,
+                    color: selectedAddress
+                        ? dark
+                        ? TColors.light
+                        : TColors.dark.withOpacity(0.2)
+                        : null,
+                  ),
+                ),
+                Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(address.name,
+                        maxLines: 1,
+                        overflow: TextOverflow.ellipsis,
+                        style: Theme.of(context).textTheme.titleLarge),
+                    const SizedBox(height: TSizes.sm / 2),
+                    Text(address.formattedPhoneNo,
+                        maxLines: 1, overflow: TextOverflow.ellipsis),
+                    const SizedBox(height: TSizes.sm / 2),
+                    Text(address.toString(),
+                        softWrap: true),
+                  ],
+                )
+              ],
             ),
           ),
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text('Phuc Le',
-                  maxLines: 1,
-                  overflow: TextOverflow.ellipsis,
-                  style: Theme.of(context).textTheme.titleLarge),
-              const SizedBox(height: TSizes.sm / 2),
-              const Text('(+84) 899 214 109',
-                  maxLines: 1, overflow: TextOverflow.ellipsis),
-              const SizedBox(height: TSizes.sm / 2),
-              const Text('252 Dang Xuan Bang, Khue Trung, Cam Le, Da Nang',
-                  softWrap: true),
-            ],
-          )
-        ],
-      ),
+        );
+      },
     );
   }
 }
