@@ -5,10 +5,11 @@ import 'package:flutter_ecommerce/common/widgets/list_tiles/settings_menu_tile.d
 import 'package:flutter_ecommerce/common/widgets/texts/section_heading.dart';
 import 'package:flutter_ecommerce/data/dummy_data.dart';
 import 'package:flutter_ecommerce/data/repositories/authentication/authentication_repository.dart';
+import 'package:flutter_ecommerce/features/shop/screens/cart/cart.dart';
+import 'package:flutter_ecommerce/features/shop/screens/order/order.dart';
 import 'package:flutter_ecommerce/data/repositories/product/product_repository.dart';
 import 'package:flutter_ecommerce/features/personalization/screens/address/address.dart';
-import 'package:flutter_ecommerce/features/shop/screens/assistant/chat_page.dart';
-import 'package:flutter_ecommerce/features/shop/screens/printer/pdf_print.dart';
+import 'package:flutter_ecommerce/features/personalization/screens/admin/load_multiple_choice.dart';
 import 'package:flutter_ecommerce/utils/constants/colors.dart';
 import 'package:flutter_ecommerce/utils/constants/sizes.dart';
 import 'package:get/get.dart';
@@ -69,13 +70,13 @@ class SettingsScreen extends StatelessWidget {
                     icon: Iconsax.shopping_cart,
                     title: 'My Cart',
                     subtitle: 'Add, remove products and move to checkout ',
-                    onTap: () {},
+                    onTap: () =>Get.to(()=> const CartScreen()),
                   ),
                   TSettingsMenuTile(
                     icon: Iconsax.bag_tick,
                     title: 'My Orders',
                     subtitle: 'In-progress and Completed Orders',
-                    onTap: () {},
+                    onTap: ()=> Get.to(()=> const OrderScreen()),
                   ),
                   TSettingsMenuTile(
                     icon: Iconsax.bank,
@@ -101,18 +102,6 @@ class SettingsScreen extends StatelessWidget {
                     subtitle: 'Manage data usage and connected accounts',
                     onTap: () {},
                   ),
-                  TSettingsMenuTile(
-                    icon: Iconsax.printer,
-                    title: 'PDF printer',
-                    subtitle: 'Print your entire data into a pdf file',
-                    onTap: () => Get.to(() => const TPdfPrinter()),
-                  ),
-                  TSettingsMenuTile(
-                    icon: Iconsax.command,
-                    title: 'Chatbot Assistant',
-                    subtitle: 'Chat with me!',
-                    onTap: () => Get.to(() => const ChatPage()),
-                  ),
 
                   ///--App Settings
                   const SizedBox(height: TSizes.spaceBtwSections),
@@ -121,9 +110,9 @@ class SettingsScreen extends StatelessWidget {
                   const SizedBox(height: TSizes.spaceBtwItems),
                   TSettingsMenuTile(
                       icon: Iconsax.document_upload,
-                      title: 'Load Data',
+                      title: 'Admin',
                       subtitle: 'Upload Data to your Cloud Firebase',
-                      onTap: () => product.uploadDummyData(TDummyData.products)),
+                      onTap: () => Get.to(() => const LoadMultipleChoice())),
                   TSettingsMenuTile(
                     icon: Iconsax.location,
                     title: 'Geolocation',
