@@ -1,6 +1,8 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_ecommerce/common/widgets/products/product_cards/product_card_vertical.dart';
 import 'package:flutter_ecommerce/common/widgets/shimmers/vertical_product_shimmer.dart';
+import 'package:flutter_ecommerce/features/shop/screens/all_products/all_products.dart';
 import 'package:flutter_ecommerce/features/shop/screens/home/widgets/home_appbar.dart';
 import 'package:flutter_ecommerce/features/shop/screens/home/widgets/home_categories.dart';
 import 'package:flutter_ecommerce/features/shop/screens/home/widgets/promo_slider.dart';
@@ -12,7 +14,7 @@ import '../../../../common/widgets/custom_shapes/containers/primary_header_conta
 import '../../../../common/widgets/custom_shapes/containers/search_container.dart';
 import '../../../../common/widgets/layouts/grid_layout.dart';
 import '../../../../common/widgets/texts/section_heading.dart';
-import '../../controllers/product_controller.dart';
+import '../../controllers/product/product_controller.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
@@ -69,7 +71,12 @@ class HomeScreen extends StatelessWidget {
                   const SizedBox(height: TSizes.spaceBtwSections),
 
                   /// -- Heading
-                  TSectionHeading(title: 'Popular products', onPressed: () {}),
+                  TSectionHeading(
+                      title: 'Popular products',
+                      onPressed: () => Get.to(() => AllProducts(
+                            title: 'Popular Products',
+                            futureMethod: controller.fetchAllFeaturedProducts(),
+                          ))),
                   const SizedBox(height: TSizes.spaceBtwItems),
 
                   /// -- Popular products
