@@ -3,20 +3,25 @@ import 'package:flutter_ecommerce/common/widgets/images/t_circular_image.dart';
 import 'package:flutter_ecommerce/common/widgets/texts/product_title_text.dart';
 import 'package:flutter_ecommerce/common/widgets/texts/t_brand_title_text_with_verified_icon.dart';
 import 'package:flutter_ecommerce/common/widgets/texts/t_product_price_text.dart';
+import 'package:flutter_ecommerce/features/shop/controllers/product/product_controller.dart';
+import 'package:flutter_ecommerce/features/shop/models/product_model.dart';
+import 'package:flutter_ecommerce/utils/enums/enums.dart';
 import 'package:flutter_ecommerce/utils/helpers/helper_functions.dart';
-import 'package:get/get.dart';
 
 import '../../../../../common/widgets/custom_shapes/containers/rounded_container.dart';
 import '../../../../../utils/constants/colors.dart';
-import '../../../../../utils/constants/enums.dart';
 import '../../../../../utils/constants/image_strings.dart';
 import '../../../../../utils/constants/sizes.dart';
 
 class TProductMetaData extends StatelessWidget {
-  const TProductMetaData({super.key});
+  const TProductMetaData({super.key, required this.product});
+
+  final ProductModel product;
 
   @override
   Widget build(BuildContext context) {
+    final controller = ProductController.instance;
+    final salePercentage = controller.calculateSalePercentage(product.price, product.salePrice);
     final darkMode = THelperFunctions.isDarkMode(context);
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
@@ -29,39 +34,41 @@ class TProductMetaData extends StatelessWidget {
                 radius: TSizes.sm,
                 backgroundColor: TColors.secondary.withOpacity(0.8),
                 padding: const EdgeInsets.symmetric(horizontal: TSizes.sm, vertical: TSizes.xs),
-                child: Text('25%', style: Theme.of(context).textTheme.labelLarge!.apply(color: TColors.black)),
+                child: Text('$salePercentage%', style: Theme.of(context).textTheme.labelLarge!.apply(color: TColors.black)),
               ),
-            SizedBox(width: TSizes.spaceBtwItems),
+            const SizedBox(width: TSizes.spaceBtwItems),
 
             /// Price
-           Text('\$46.67', style: Theme.of(context).textTheme.titleSmall!.apply(decoration: TextDecoration.lineThrough)),
-            const SizedBox(width: TSizes.spaceBtwItems),
-            TProductPriceText(price: '35',isLarge: true),
+            if (product.productType == ProductType.single.toString() && product.salePrice > 0)
+              Text('\$${product.price}', style: Theme.of(context).textTheme.titleSmall!.apply(decoration: TextDecoration.lineThrough)),
+            if (product.productType == ProductType.single.toString() && product.salePrice > 0) const SizedBox(width: TSizes.spaceBtwItems),
+            TProductPriceText(price: controller.getProductPrice(product),isLarge: true),
           ],
         ),
-        SizedBox(height: TSizes.spaceBtwItems/1.5),
+        const SizedBox(height: TSizes.spaceBtwItems/1.5),
         /// Title
-         TProductTitleText(title:'Nike Air Force 1'),
+         TProductTitleText(title: product.title),
          SizedBox(height: TSizes.spaceBtwItems/1.5),
         /// Stock Status
         Row(
           children: [
             const TProductTitleText(title:'Status:'),
             SizedBox(width: TSizes.spaceBtwItems),
-            Text('In Stock', style: Theme.of(context).textTheme.titleMedium),
+            Text(controller.getProductStockStatus(product.stock), style: Theme.of(context).textTheme.titleMedium),
           ],
         ),
-        SizedBox(height: TSizes.spaceBtwItems/1.5),
+        const SizedBox(height: TSizes.spaceBtwItems/1.5),
         /// Brand
         Row(
           children: [
             TCircularImage(
-                image: TImages.shoeIcon,
+              isNetworkImage: true,
+              image: TImages.nikeLogo,
               width: 32,
               height: 32,
               overlayColor: darkMode ? TColors.white: TColors.black,
             ),
-            const TBrandTitleWithVerifiedIcon(title: 'Nike', brandTextSize: TextSizes.medium, )
+            TBrandTitleWithVerifiedIcon(title: product.brand != null ? product.brand!.name : '')
           ],
         ),
       ],
