@@ -1,112 +1,114 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_ecommerce/common/widgets/appbar/appbar.dart';
 import 'package:flutter_ecommerce/common/widgets/appbar/tabbar.dart';
+import 'package:flutter_ecommerce/common/widgets/brands/t_brand_card.dart';
 import 'package:flutter_ecommerce/common/widgets/custom_shapes/containers/rounded_container.dart';
 import 'package:flutter_ecommerce/common/widgets/custom_shapes/containers/search_container.dart';
 import 'package:flutter_ecommerce/common/widgets/layouts/grid_layout.dart';
 import 'package:flutter_ecommerce/common/widgets/products/cart/cart_menu_icon.dart';
 import 'package:flutter_ecommerce/common/widgets/texts/section_heading.dart';
 import 'package:flutter_ecommerce/common/widgets/texts/t_brand_title_text_with_verified_icon.dart';
+import 'package:flutter_ecommerce/features/shop/controllers/brand_controller.dart';
 import 'package:flutter_ecommerce/features/shop/controllers/category_controller.dart';
+import 'package:flutter_ecommerce/features/shop/screens/brand/all_brands.dart';
+import 'package:flutter_ecommerce/features/shop/screens/brand/brand_products.dart';
 import 'package:flutter_ecommerce/features/shop/screens/store/widgets/category_tab.dart';
 import 'package:flutter_ecommerce/utils/constants/colors.dart';
 import 'package:flutter_ecommerce/utils/constants/enums.dart';
 import 'package:flutter_ecommerce/utils/constants/image_strings.dart';
 import 'package:flutter_ecommerce/utils/constants/sizes.dart';
 import 'package:flutter_ecommerce/utils/helpers/helper_functions.dart';
+import 'package:get/get.dart';
 
 import '../../../../common/widgets/images/t_circular_image.dart';
+import '../../../../common/widgets/shimmers/brands_shimmer.dart';
 
 class StoreScreen extends StatelessWidget {
   const StoreScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
+    final brandController = Get.put(BrandController());
     final categories = CategoryController.instance.featuredCategories;
     return DefaultTabController(
       length: categories.length,
       child: Scaffold(
         appBar: TAppBar(
-          title: Text('Store', style: Theme.of(context).textTheme.headlineMedium),
+          title:
+              Text('Store', style: Theme.of(context).textTheme.headlineMedium),
           actions: [
-            TCartCounterIcon(onPressed: (){}, iconColor: TColors.black)
+            TCartCounterIcon(onPressed: () {}, iconColor: TColors.black)
           ],
         ),
-        body: NestedScrollView(headerSliverBuilder: (_, innerBoxIsScrolled) {
-          return [
-            SliverAppBar(
-              automaticallyImplyLeading: false,
-              pinned: true,
-              floating: true,
-              backgroundColor: THelperFunctions.isDarkMode(context) ? TColors.black : TColors.white,
-              expandedHeight: 440,
+        body: NestedScrollView(
+            headerSliverBuilder: (_, innerBoxIsScrolled) {
+              return [
+                SliverAppBar(
+                  automaticallyImplyLeading: false,
+                  pinned: true,
+                  floating: true,
+                  backgroundColor: THelperFunctions.isDarkMode(context)
+                      ? TColors.black
+                      : TColors.white,
+                  expandedHeight: 440,
 
-              flexibleSpace: Padding(
-                  padding: const EdgeInsets.all(TSizes.defaultSpace),
-                  child: ListView(
-                    shrinkWrap: true,
-                    physics: const NeverScrollableScrollPhysics(),
-                    children: [
-                      /// -- Search bar
-                      const SizedBox(height: TSizes.spaceBtwItems),
-                      const TSearchContainer(text: 'Search in store', showBorder: true, showBackground: false, padding: EdgeInsets.zero),
-                      const SizedBox(height: TSizes.spaceBtwSections),
-
-                      /// -- Featured Brands
-                      TSectionHeading(title: 'Featured Brands', onPressed: (){}),
-                      const SizedBox(height: TSizes.spaceBtwItems / 1.5),
-
-                      TGridLayout(itemCount: 4, maxAxisExtent: 80, itemBuilder: (_, index){
-                        return GestureDetector(
-                          onTap: (){},
-                          child: TRoudedContainer(
-                            padding: const EdgeInsets.all(TSizes.sm),
+                  flexibleSpace: Padding(
+                    padding: const EdgeInsets.all(TSizes.defaultSpace),
+                    child: ListView(
+                      shrinkWrap: true,
+                      physics: const NeverScrollableScrollPhysics(),
+                      children: [
+                        /// -- Search bar
+                        const SizedBox(height: TSizes.spaceBtwItems),
+                        const TSearchContainer(
+                            text: 'Search in store',
                             showBorder: true,
-                            backgroundColor: Colors.transparent,
-                            child: Row(
-                              children: [
-                                /// -- Icon
-                                Flexible(
-                                  child: TCircularImage(
-                                    isNetworkImage: false,
-                                    image: TImages.clothIcon,
-                                    backgroundColor: Colors.transparent,
-                                    overlayColor: THelperFunctions.isDarkMode(context) ? TColors.white : TColors.black,
-                                  ),
-                                ),
-                                const SizedBox(width: TSizes.spaceBtwItems / 2),
+                            showBackground: false,
+                            padding: EdgeInsets.zero),
+                        const SizedBox(height: TSizes.spaceBtwSections),
 
-                                /// -- Text
-                                Expanded(
-                                  child: Column(
-                                    mainAxisSize: MainAxisSize.min,
-                                    crossAxisAlignment: CrossAxisAlignment.start,
-                                    children: [
-                                      const TBrandTitleWithVerifiedIcon(title: 'NIKE', brandTextSize: TextSizes.large),
-                                      Text(
-                                        '256 products',
-                                        overflow: TextOverflow.ellipsis,
-                                        style: Theme.of(context).textTheme.labelMedium,
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                              ],
-                            ),
-                          ),
-                        );
-                      })
-                    ],
+                        /// -- Featured Brands
+                        TSectionHeading(
+                            title: 'Featured Brands',
+                            onPressed: () =>
+                                Get.to(() => const AllBrandsScreen())),
+                        const SizedBox(height: TSizes.spaceBtwItems / 1.5),
+
+                        /// brand grid
+                        Obx(() {
+                          if (brandController.isLoading.value) return const TBrandsShimmer();
+
+                          if (brandController.featuredBrands.isEmpty) {
+                            return Center(
+                              child: Text('No Data Found!', style: Theme.of(context).textTheme.bodyMedium!.apply(color: Colors.white)),
+                            );
+                          }
+
+                          return TGridLayout(
+                              itemCount: brandController.featuredBrands.length,
+                              maxAxisExtent: 80,
+                              itemBuilder: (_, index) {
+                                final brand = brandController.featuredBrands[index];
+                                return TBrandCard(showBorder: true, brand: brand, onTap: () => Get.to(() => BrandProducts(brand: brand)),);
+                              });
+                        }),
+                      ],
+                    ),
                   ),
-              ),
 
-              /// Tabs
-              bottom: TTabBar(
-                  tabs: categories.map((category) =>Tab(child: Text(category.name))).toList(),
-              ),
-            ),
-          ];
-        }, body: TabBarView(children: categories.map((category) => TCategorytab(category: category)).toList())),
+                  /// Tabs
+                  bottom: TTabBar(
+                    tabs: categories
+                        .map((category) => Tab(child: Text(category.name)))
+                        .toList(),
+                  ),
+                ),
+              ];
+            },
+            body: TabBarView(
+                children: categories
+                    .map((category) => TCategorytab(category: category))
+                    .toList())),
       ),
     );
   }
